Fixed MIDI devices on iOS
The default browser on iOS (Safari) does not have MIDI device support.
Most browsers (Chrome, Opera) available on the store use the Safari
engine internally, and will therefore also not support MIDI devices.

There is, however, an app called Web MIDI, that is available from
https://itunes.apple.com/us/app/web-midi-browser/id953846217?mt=8
This app can load a web page on iOS with MIDI device support.

Unfortunately, this app does not seem to support using the spread
operator to convert an interator to an array. This commit changes these
to regular for loops, which do work.